pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pretty-qr"
version = "0.1.0"
description = "Stylized rounded QR code generator"
requires-python = ">=3.10"
dependencies = [
  "Pillow>=10.0.0",
  "qrcode>=7.4.2",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[tool.pytest.ini_options]
pythonpath = ["."]
